STL error fix for MinGW
When namespace std is using namespace tr1, the MinGW STL generates a
compile error due to an ambiguity between the std::__detail and the
std::tr1::__detail namespaces (internal to the implementation).  By
changing stl to use both std and std::tr1, the std namespace remains
unchanged and the ambiguity is avoided.