Added flag to control building examples.
This PR adds a new flag to `genie.lua` called `--with-examples`
and uses it to control whether the example binaries are built or
not.

It also changes the default options in `makefile` to include this
flag, such that the default behavior of `bgfx` compilation is
unchanged.

This allows users who are building with CI systems or as part of a
larger automated build process to avoid having to compile examples
each time they build `bgfx`.  They can instead use `genie` directly
to configure their project to build without examples.