Added BGFX_CONFIG_D3D11_CREATE_DEVICE_MULTITHREADED
In the use case this is intended for, I am re-using the D3D11 device created by BGFX for hardware accelerated video decoding using the FFmpeg libraries. This is only possible when the D3D11 device has been created with multithreaded protection and is better than having multiple D3D11 devices and having to deal with shared texture resources etc.