Allow Metal to work when OpenGL is turned off.
If BGFX_CONFIG_RENDERER_OPENGL{ES} is set to 0, in order to remove the
OpenGL codebase, the Metal renderer is not seen as supported.
The issue is in the initialisation of s_rendererCreator[Meta].supported
as BGFX_CONFIG_RENDERER_METAL defaults to 0.  This change sets that
value to 1 for iOS/macOS which makes it work again.  It's not perfect
though as Vulkan support also disappears, but that doesn't matter on
iOS/macOS so I've left that as is.