Don't encode invalid programs as valid unrelated programs
When an invalid program is submitted with bgfx::touch(), it gets encoded
as program index 0 by the command encoder. Since program 0 can be a
valid program (typically it's the debug text blitter program), this can
result in garbage being drawn if no other rendering is submitted for the
same view.

This patch changes invalid programs to be encoded as invalid programs,
so the renderer can properly ignore them.